renamed package department.entity to entity
