+++ src/main/java/se/jocke/entity/Employee.java
@@ -1,4 +1,4 @@
-package se.jocke.department.entity;
+package se.jocke.entity;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
