+++ src/main/java/se/jocke/service/EmployeeService.java
@@ -1,6 +1,6 @@
 package se.jocke.service;
 
-import se.jocke.department.entity.Employee;
+import se.jocke.entity.Employee;
 
 import java.util.List;
