+++ src/main/java/se/jocke/service/EmployeeServiceImpl.java
@@ -7,7 +7,7 @@ import se.jocke.dao.EmployeeDao;
 import se.jocke.dao.EmployeeDatabaseEntry;
 import se.jocke.dao.EntityNotFoundException;
 import se.jocke.dao.mapper.EmployeePojoMapper;
-import se.jocke.department.entity.Employee;
+import se.jocke.entity.Employee;
 
 import java.util.List;
 import java.util.Optional;
