+++ src/test/java/se/jocke/employee/api/model/TestCreateEmployeeModel.java
@@ -12,6 +12,8 @@ public class TestCreateEmployeeModel {
 
     @Test
     public void testCreateEmployeeModel() {
+        //EmployeeModel emp = new EmployeeModel();
+
         EmployeeModel employeeModel = EmployeeModel.builder()
                 .employeeId(EMPLOYEE_MODEL.getEmployeeId())
                 .firstName(EMPLOYEE_MODEL.getFirstName())
