+++ src/test/java/se/jocke/employee/entity/TestCreateEmployee.java
@@ -2,7 +2,7 @@ package se.jocke.employee.entity;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import se.jocke.department.entity.Employee;
+import se.jocke.entity.Employee;
 import se.jocke.employee.builder.EmployeeTestBuilder;
 
 public class TestCreateEmployee {
@@ -10,10 +10,10 @@ public class TestCreateEmployee {
 
     @Test
     public void testCreateEmployee() {
-        Employee employee = EmployeeTestBuilder.builder()
+        Employee employee = Employee.builder()
                 .employeeId(EMPLOYEE.getEmployeeId())
                 .firstName(EMPLOYEE.getFirstName())
-                .lastName(EMPLOYEE.getLastName())
+                //.lastName(EMPLOYEE.getLastName())
                 .fullTime(EMPLOYEE.getFullTime())
                 .salary(EMPLOYEE.getSalary())
                 .departmentId(EMPLOYEE.getDepartmentId())
