+++ src/test/java/se/jocke/employee/service/TestEmployeeService.java
@@ -12,7 +12,7 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
 import se.jocke.dao.EmployeeDao;
 import se.jocke.dao.EmployeeDatabaseEntry;
-import se.jocke.department.entity.Employee;
+import se.jocke.entity.Employee;
 import se.jocke.service.EmployeeService;
 import se.jocke.service.EmployeeServiceImpl;
